+++ libs/request/src/CommonMediaRequest.ts
@@ -1,9 +1,9 @@
 import type { Cmcd } from '@svta/cml-cmcd'
-import type { Request } from '@svta/cml-utils'
+import type { HttpRequest } from '@svta/cml-utils'
 
 /**
  * Common Media Request.
  *
  * @public
  */
-export type CommonMediaRequest = Request<{ cmcd: Cmcd }>;
+export type CommonMediaRequest = HttpRequest<{ cmcd: Cmcd }>;