+++ libs/request/src/RequestResponseType.ts
@@ -1,0 +1,6 @@
+/**
+ * The response type of the request.
+ *
+ * @public
+ */
+export { RequestResponseType } from '@svta/cml-utils'