+++ libs/utils/src/Request.ts
@@ -1,55 +1,9 @@
-import type { RequestType } from './RequestType.ts'
+import type { HttpRequest } from './HttpRequest.ts'
 
 /**
- * Generic request API.
+ * Request API.
  *
  * @public
+ * @deprecated Use {@link HttpRequest} instead.
  */
-export type Request<D = any> = {
-
-	/**
-	 * The URL of the request.
-	 */
-	url: string;
-
-	/**
-	 * The request's method (GET, POST, etc).
-	 */
-	method?: string;
-
-	/**
-	 * The body of the request.
-	 */
-	body?: BodyInit;
-
-	/**
-	 * The response type with which the response from the server shall be compatible.
-	 */
-	responseType?: RequestType;
-
-	/**
-	 * The headers object associated with the request.
-	 */
-	headers?: Record<string, string>;
-
-	/**
-	 * Indicates whether the user agent should send or receive cookies from the other domain in the case of cross-origin requests.
-	 */
-	credentials?: RequestCredentials;
-
-	/**
-	 * The mode of the request (e.g., cors, no-cors, same-origin, etc).
-	 */
-	mode?: RequestMode;
-
-	/**
-	 * The number of milliseconds the request can take before automatically being terminated.
-	 * If undefined or value is 0 then there is no timeout.
-	 */
-	timeout?: number;
-
-	/**
-	 * Any custom data.
-	 */
-	customData?: D;
-};
+export type Request = HttpRequest